refactor(ui): extract code cell painting
